backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
torch
torchvision
numpy
opencv-python-headless
pillow
pydantic
fpdf2
